requirements.txt
# Django and Core Dependencies
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1

# Core Python packages
setuptools>=68.0.0

# Authentication and JWT
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# Database
psycopg2-binary==2.9.9
PyMySQL==1.1.0

# Environment and Configuration
python-decouple==3.8
python-dotenv==1.0.1

# Password Hashing
argon2-cffi==23.1.0
bcrypt==4.1.2

# API Documentation
drf-spectacular==0.27.2
drf-spectacular[sidecar]==0.27.2

# Rate Limiting
django-ratelimit==4.1.0

# Validation
pydantic==2.7.1

# Development and Testing
pytest==8.2.0
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0

# Logging and Monitoring
structlog==24.1.0
django-extensions==3.2.3

# WSGI Server
gunicorn==22.0.0

# Development Tools
black==24.4.0
isort==5.13.2
flake8==7.0.0
mypy==1.10.0
django-stubs==5.0.0

# Utilities
celery==5.3.6
redis==5.0.4
